backend/pyproject.toml
[project]
name = "source-comparison-backend"
version = "0.1.0"
description = "FastAPI backend for a DeepSeek-powered sourcing & price comparison agent."
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy[asyncio]>=2.0.30",
  "asyncmy>=0.2.9",
  "pydantic>=2.7.0",
  "python-multipart>=0.0.9",
  "python-dotenv>=1.0.1",
  "httpx>=0.27.0",
  "openai>=1.40.0",
  "cryptography",
  "pandas>=2.2.0",
  "openpyxl>=3.1.0",
  "python-docx>=1.1.0",
  "pypdf>=4.0.0",
  "python-pptx>=0.6.23",
]

[project.optional-dependencies]
dev = [
  "ipython",
  "ruff",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
  "ipython",
  "ruff",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]
